# pkg/objectfile: do not cache base