even same mapping file.
The base address will changed when addr changed, so we should not cache it.
```
benchmark                  old ns/op     new ns/op     delta
BenchmarkELFObjAddr-10     2.47          8.65          +249.62%
```