@kakkoyun 2023-05-09
Deprecated after #1507.

Thanks a lot for the work ❤️ 
We're still open to improving the mechanism around this. Let us know if you need something specific.